refactor(module.json): Update module name
After some discussion with other module developers, the general
consensus is that you should avoid certain worlds like "system" in your
module if possible, as they are usually refering to other things in the
Foundry ecosystem. With this we are dropping the system and making the
module name "tabulanil-sanity-dnd5e", more explicit about the system
compatibility and more clear about it's purpose.