feat(tabulanil-sanity.js): add readonly UI element
Now a "Sanity bar" is shown below the Hit Dice information on the Actor
Sheet from dnd53 v3. We are basing this element on the Hit Point
element, so it uses the same style, but some of the functions are
changed. The bar will show current/maximum amount of sanity points, and
the field that would show the temporary hit points will show the current
sanity tier for that charcter.

TODO: on Hover display tooltip with effects