fix(tabulanil-sanity.js): undefined flag init
The module flags where not being correctly initialized as they depend
on the `update` from the document, and that makes async calls
internally, meaning that some of the code was dealing with undefined or
null values. The new approach creates an opject that updates the flags
in a single call.

This will trigger a refactor of the module's utility methods as they
should now return objects and values and the update happen only in a
single call.