refactor(tabulanil-sanity.js): remove helpers
With the new approach trying to simplify the module we are no longer
needing to deal with too many flag updates, which means that we only
need to handle a single persistent flag for the current sanity points.

That made so part of the helper methods could be removed or rewritten to
calculate their values on the go, reducing the overall data storage
needed.