fix(tabulanil-sanity.js): incorrect flag checking
Previously, the implementation would check if there was a sanity flag on
the actor, if not it would initialize the module flags properly. When a
actor reached 0 sanity, the result of that comparison, relying on the
fact that the flag would have only positive, non null values would then
become true, and make initialization happen again. This resulted in a
bug where the actor could never have its sanity equal 0 (zero).

To fix this, we changed the check to rely on the `undefined` value as if
the flag is not set, the result would be an undefined value. We also
added a "clamping" mechanism so the current sanity cannot go beyond its
maximum (total sanity) and below 0 (character is effectively dead).

Some extra log messages were added to help debug issues with this new
implementation.