feat: report Insanity Tier changes in chat
Now the when the change in sanity of an actor changes the insanity tier
of that actor a chat message will be sent as a whisper to all the owners
of that actor (including the GM). This has been done so the players and
GM do not need to keep looking in the actor sheet if there was a change
on the insanity tier.