chore(module.json): compatibility with v12
Module has been verified to work on v12 without issues

docs(readme.md): update version information

fix(tabulanil-sanity.js): update TokenHUD UI state

When clamping values, on the case of a positive overflow, the input
element on the TokenHUD would keep the input value until an unfocus
event happened. An extra line is being added to match the value of the
input element to the clamped value.

There's a possibility this is not a real issue, but to ensure consistent
UI state this has been deemed necessary for now.

closes #2